+++ Craftsman/Builders/Dtos/DtoFileTextGenerator.cs
@@ -3,6 +3,7 @@
     using Craftsman.Enums;
     using Craftsman.Helpers;
     using Craftsman.Models;
+    using Craftsman.Models.Interfaces;
     using System;
     using System.Collections.Generic;
     using System.Linq;
@@ -25,13 +26,22 @@
 }}";
         }
 
-        public static string GetDtoText(ClassPath dtoClassPath, Entity entity, Dto dto, string projectBaseName)
+        public static string GetDtoText(IClassPath dtoClassPath, Entity entity, Dto dto, string projectBaseName)
         {
-            var propString = dto == Dto.Creation || dto == Dto.Update ? "" : DtoPropBuilder(entity.Properties, dto);
-            var abstractString = dto == Dto.Manipulation ? $"abstract" : "";
+            var propString = DtoPropBuilder(entity.Properties, dto);
+            if (dto == Dto.Creation)
+            {
+                propString = "";
+                if (entity.PrimaryKeyProperty.Type.Equals("guid", StringComparison.InvariantCultureIgnoreCase))
+                    propString = DtoPropBuilder(new List<EntityProperty>() { entity.PrimaryKeyProperty }, dto);
+            }
+            else if (dto == Dto.Update)
+                propString = "";
+
+            var abstractString = dto == Dto.Manipulation ? $"abstract " : "";
 
             var inheritanceString = "";
-            if(dto == Dto.Creation || dto == Dto.Update)
+            if (dto == Dto.Creation || dto == Dto.Update)
                 inheritanceString = $": {Utilities.GetDtoName(entity.Name, Dto.Manipulation)}";
 
             return @$"namespace {dtoClassPath.ClassNamespace}
@@ -40,9 +50,9 @@
     using System.ComponentModel.DataAnnotations;
     using System.ComponentModel.DataAnnotations.Schema;
 
-    public {abstractString} class {Utilities.GetDtoName(entity.Name, dto)} {inheritanceString}
+    public {abstractString}class {Utilities.GetDtoName(entity.Name, dto)} {inheritanceString}
     {{
-{propString}
+{propString.RemoveLastNewLine()}
 
         // add-on property marker - Do Not Delete This Comment
     }}
@@ -54,21 +64,24 @@
             var dtoFileName = $"{Utilities.GetDtoName(prop.Type, dto)}.cs";
             var fkClasspath = ClassPathHelper.DtoClassPath(dtoClassPath.SolutionDirectory, dtoFileName, prop.Type, projectBaseName);
 
-            fkUsingStatements += $"{Environment.NewLine}    using {fkClasspath.ClassNamespace};";   
-            
+            fkUsingStatements += $"{Environment.NewLine}    using {fkClasspath.ClassNamespace};";
+
             return fkUsingStatements;
         }
 
         public static string DtoPropBuilder(List<EntityProperty> props, Dto dto)
         {
             var propString = "";
-            for(var eachProp = 0; eachProp < props.Count; eachProp++)
+            for (var eachProp = 0; eachProp < props.Count; eachProp++)
             {
                 if (!props[eachProp].CanManipulate && dto == Dto.Manipulation)
                     continue;
+                var guidDefault = dto == Dto.Creation && props[eachProp].Type.Equals("guid", StringComparison.InvariantCultureIgnoreCase)
+                    ? " = Guid.NewGuid();"
+                    : "";
 
                 string newLine = eachProp == props.Count - 1 ? "" : Environment.NewLine;
-                propString += $@"        public {props[eachProp].Type} {props[eachProp].Name} {{ get; set; }}{newLine}";
+                propString += $@"        public {props[eachProp].Type} {props[eachProp].Name} {{ get; set; }}{guidDefault}{newLine}";
             }
 
             return propString;